# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Droppy",
        "title": "Droppy API",
        "contact": {
            "name": "Droppy API Support",
            "email": "stervinou.g36@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/comments/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drop"
                ],
                "summary": "Delete a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "No"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errors2.MultiFieldsError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/comments/{id}/responses": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Respond to a comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment response"
                ],
                "summary": "Respond to a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment creation object",
                        "name": "response",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CommentCreationParam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response_models.GetCommentResponseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errors2.MultiFieldsError"
                        }
                    }
                }
            }
        },
        "/comments/{id}/responses/{responseId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a comment response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment response"
                ],
                "summary": "Delete a comment response",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Comment response ID",
                        "name": "responseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": ""
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errors2.MultiFieldsError"
                        }
                    }
                }
            }
        },
...
        "/drops/{id}/comments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Comment on a drop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drop"
                ],
                "summary": "Comment on a drop",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Drop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment creation object",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CommentCreationParam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response_models.GetCommentResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errors2.MultiFieldsError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errors2.MultiFieldsError": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.CommentCreationParam": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
...
        "response_models.GetCommentResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "createdBy": {},
                "drop": {
                    "$ref": "#/definitions/response_models.GetDropResponse"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "response_models.GetCommentResponseForDrop": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "createdBy": {},
                "id": {
                    "type": "integer"
                },
                "responses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response_models.GetCommentResponseResponse"
                    }
                }
            }
        },
        "response_models.GetCommentResponseResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "createdBy": {},
                "id": {
                    "type": "integer"
                }
            }
        },
        "response_models.GetDropResponse": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response_models.GetCommentResponseForDrop"
                    }
                },
                "content": {
                    "type": "string"
                },
                "contentPicturePath": {
                    "type": "string"
                },
                "contentSubtitle": {
                    "type": "string"
                },
                "contentTitle": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {},
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isCurrentUserLiking": {
                    "type": "boolean"
                },
                "isPinned": {
                    "type": "boolean"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "location": {
                    "type": "string"
                },
                "picturePath": {
                    "type": "string"
                },
                "totalComments": {
                    "type": "integer"
                },
                "totalLikes": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}